test: fix broken build after refactoring. TODO: More archunit fixes
