+++ use-core/pom.xml
@@ -82,6 +82,18 @@
     <build>
         <plugins>
             <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-jar-plugin</artifactId>
+                <version>3.3.0</version>
+                <executions>
+                    <execution>
+                        <goals>
+                            <goal>test-jar</goal>
+                        </goals>
+                    </execution>
+                </executions>
+            </plugin>
+            <plugin>
                 <groupId>com.bekioui.maven.plugin</groupId>
                 <artifactId>merge-maven-plugin</artifactId>
                 <version>1.2.0</version>
