+++ use-core/src/main/java/module-info.java
@@ -39,6 +39,7 @@ module use.core {
     exports org.tzi.use.gen.tool;
     exports org.tzi.use.parser.shell;
     exports org.tzi.use.parser.testsuite;
+    exports org.tzi.use.core.test;
     exports org.tzi.use.api;
     exports org.tzi.use.main;
     exports org.tzi.use.parser.generator;
