+++ use-core/src/main/java/org/tzi/use/core/test/TestModelUtil.java
@@ -17,10 +17,12 @@
  * Foundation, Inc., 675 Mass Ave, Cambridge, MA 02139, USA.
  */
 
-package org.tzi.use.uml.mm;
+package org.tzi.use.core.test;
 
 import org.tzi.use.api.UseApiException;
 import org.tzi.use.api.UseModelApi;
+import org.tzi.use.uml.mm.MModel;
+import org.tzi.use.uml.mm.MAggregationKind;
 
 import java.util.ArrayList;
 import java.util.List;
