+++ use-gui/src/test/java/org/tzi/use/gui/views/diagrams/util/CreationTimeRecorderTest.java
@@ -28,7 +28,7 @@ import org.tzi.use.api.UseSystemApi;
 import org.tzi.use.gui.views.diagrams.behavior.communicationdiagram.CreationTimeRecorder;
 import org.tzi.use.gui.views.diagrams.behavior.communicationdiagram.MMessage;
 import org.tzi.use.uml.mm.MModel;
-import org.tzi.use.uml.mm.TestModelUtil;
+import org.tzi.use.core.test.TestModelUtil;
 import org.tzi.use.uml.sys.MObject;
 import org.tzi.use.uml.sys.MSystem;
 import org.tzi.use.uml.sys.events.Event;
